refactor: discount관련 컴포넌트 폴더 구분
