--- src/refactoring/components/Admin/product/ProductItem.tsx
@@ -1,14 +1,0 @@
-import { Discount } from "../../../../types";
-
-interface Props {
-  discount: Discount;
-}
-export const ProductItem = ({ discount }: Props) => {
-  return (
-    <div className="mb-2">
-      <span>
-        {discount.quantity}개 이상 구매 시 {discount.rate * 100}% 할인
-      </span>
-    </div>
-  );
-};
